pyproject.toml
[project]
name = "dbchat-mysql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "groq>=0.37.0",
    "langchain>=1.1.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-groq>=1.1.0",
    "mysql-connector-python==8.3.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.0",
]
